expand build_c_library to build_c_artifact
