+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["benchmark", "core", "tools/full_source", "tools/build_project_spec", "tools/load_raw_source", "tools/raw_source_to_cargo_llm", "tools/try_cargo_build", "tools/write_output", "translate", "tools/c_ast", "tools/modular_translation_llm", "tools/quantize_rust_spans", "tools/fix_declarations_llm", "tools/translate_agentic", "tools/verify_fix_agentic", "tools/generate_difftest_suite", "tools/build_c_library", "tools/run_difftest", "tools/fix_diff_failures"]
+members = ["benchmark", "core", "tools/full_source", "tools/build_project_spec", "tools/load_raw_source", "tools/raw_source_to_cargo_llm", "tools/try_cargo_build", "tools/write_output", "translate", "tools/c_ast", "tools/modular_translation_llm", "tools/quantize_rust_spans", "tools/fix_declarations_llm", "tools/translate_agentic", "tools/verify_fix_agentic", "tools/generate_difftest_suite", "tools/build_c_artifact", "tools/run_difftest", "tools/fix_diff_failures"]
 resolver = "3"
 
 [workspace.dependencies]
@@ -19,7 +19,7 @@ fix_declarations_llm = { path = "tools/fix_declarations_llm" }
 translate_agentic = { path = "tools/translate_agentic" }
 verify_fix_agentic = { path = "tools/verify_fix_agentic" }
 generate_difftest_suite = { path = "tools/generate_difftest_suite" }
-build_c_library = { path = "tools/build_c_library" }
+build_c_artifact = { path = "tools/build_c_artifact" }
 run_difftest = { path = "tools/run_difftest" }
 fix_diff_failures = { path = "tools/fix_diff_failures" }
 log = "0.4.28"
