+++ tools/run_difftest/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-build_c_library.workspace = true
+build_c_artifact.workspace = true
 full_source.workspace = true
 generate_difftest_suite.workspace = true
 harvest_core.workspace = true
